Use GitHub API v3 to import bugs and comments
Use GitHub API v3 to import bugs and comments.  This advantage of v3
of the API is that comments are shown as proper comments in GitHub
with its original date.

The disadvantage of v3 of the API is that it doesn't allow you to
update existing issues, so this functionality is lost.

This patch is based on a patch from Théo Zimmermann of the Coq project,
see https://www.theozimmermann.net/2017/10/bugzilla-to-github/